+++ src/_pytask/collect.py
@@ -91,7 +91,7 @@ def pytask_collect_file_protocol(session, path, reports):
 
 @hookimpl
 def pytask_collect_file(session, path, reports):
-    if path.name.startswith("task_") and path.suffix == ".py":
+    if any(path.match(pattern) for pattern in session.config["task_files"]):
         spec = importlib.util.spec_from_file_location(path.stem, str(path))
 
         if spec is None:
@@ -128,6 +128,7 @@ def pytask_collect_task_protocol(session, reports, path, name, obj):
         task = session.hook.pytask_collect_task(
             session=session, path=path, name=name, obj=obj
         )
+        session.hook.pytask_collect_task_teardown(session=session, task=task)
         if task is not None:
             return CollectionReportTask.from_task(task)