+++ docs/changes.rst
@@ -6,7 +6,15 @@ chronological order. Releases follow `semantic versioning <https://semver.org/>`
 all releases are available on `Anaconda.org <https://anaconda.org/pytask/pytask>`_.
 
 
-0.0.6 - 2020-xx-xx
+0.0.7 - 2020-xx-xx
+------------------
+
+- :gh:`25` allows to customize the names of the task files.
+- :gh:`27` implements the ``pytask_collect_task_teardown`` hook specification to perform
+  checks after a task is collected.
+
+
+0.0.6 - 2020-09-12
 ------------------
 
 - :gh:`16` reduces the traceback generated from tasks, failure section in report, fix